Move `cli` to `internal/` packages
