+++ internal/cli/command_query.go
